minor reorganising of app dev features
