+++ TIMES-NZ-INTERNAL-QA/src/times_nz_internal_qa/app/app_module_developers.py
@@ -1,5 +1,5 @@
 """
-App processing for dummy processes
+App processing for developer-facing QA views
 """
 
 from functools import lru_cache
@@ -9,6 +9,7 @@ import polars as pl
 from shiny import reactive
 from times_nz_internal_qa.app.helpers.data_processing import (
     aggregate_by_group,
+    filter_df_for_variable,
     read_data_pl,
 )
 from times_nz_internal_qa.app.helpers.filters import (
@@ -25,12 +26,13 @@ from times_nz_internal_qa.utilities.filepaths import FINAL_DATA
 
 # CONSTANTS ---------------------------------------
 
-ID_PREFIX = "dum"
+ID_PREFIX = "dev"
 
 
 CLEAN_RESULTS = Path(FINAL_DATA).parent / "clean_results"
 DUMMY_ENERGY_FILEPATH = CLEAN_RESULTS / "dummy_energy.parquet"
 DUMMY_DEMAND_FILEPATH = CLEAN_RESULTS / "dummy_demand.parquet"
+TECHNOLOGY_CAPACITY_FILE = FINAL_DATA / "technology_capacity.parquet"
 
 dummy_energy_group_options = ["Fuel", "Commodity", "Region"]
 dummy_demand_group_options = [
@@ -38,6 +40,15 @@ dummy_demand_group_options = [
     "Sector",
     "EndUse",
     "Commodity",
+    "Region",
+]
+technology_capacity_group_options = [
+    "SectorGroup",
+    "Sector",
+    "TechnologyGroup",
+    "Technology",
+    "EnduseGroup",
+    "EndUse",
     "Region",
 ]
 
@@ -65,15 +76,29 @@ dummy_energy_filters = create_filter_dict(
     ],
 )
 
+technology_capacity_filters = create_filter_dict(
+    "technology_capacity",
+    [
+        {"col": "SectorGroup"},
+        {"col": "Sector"},
+        {"col": "TechnologyGroup"},
+        {"col": "Technology"},
+        {"col": "EnduseGroup"},
+        {"col": "EndUse"},
+        {"col": "Region"},
+    ],
+)
+
 
 # Specific data processing
 
 base_cols = ["Scenario", "Period", "Variable", "Unit"]
+technology_capacity_all_group_options = base_cols + technology_capacity_group_options
 
 
 # GET MAIN DATA
 @lru_cache(maxsize=8)
-def get_dum_energy_df(scenarios, filepath=DUMMY_ENERGY_FILEPATH):
+def get_dev_energy_df(scenarios, filepath=DUMMY_ENERGY_FILEPATH):
     """
     standard
     """
@@ -87,7 +112,7 @@ def get_dum_energy_df(scenarios, filepath=DUMMY_ENERGY_FILEPATH):
 
 
 @lru_cache(maxsize=8)
-def get_dum_demand_df(scenarios, filepath=DUMMY_DEMAND_FILEPATH):
+def get_dev_demand_df(scenarios, filepath=DUMMY_DEMAND_FILEPATH):
     """
     standard
     """
@@ -100,30 +125,53 @@ def get_dum_demand_df(scenarios, filepath=DUMMY_DEMAND_FILEPATH):
     return df.collect()
 
 
-dummy_energy_parameters = {
+@lru_cache(maxsize=8)
+def get_base_technology_capacity_df(scenarios, filepath=TECHNOLOGY_CAPACITY_FILE):
+    """
+    Returns non-transport technology capacity data (pre-filtered)
+    Based on scenario selections
+    Caches results for quick switching
+    """
+    df = read_data_pl(filepath, scenarios)
+    df = aggregate_by_group(df, technology_capacity_all_group_options)
+    df = filter_df_for_variable(df, "Technology Capacity", collect=True)
+    return df
+
+
+developer_energy_parameters = {
     "page_id": ID_PREFIX,
-    "chart_id": "dum_nrg",
-    "sec_id": "dum-nrg",
+    "chart_id": "dev_nrg",
+    "sec_id": "dev-nrg",
     "filters": dummy_energy_filters,
     "section_title": "Infeasible energy",
     "base_cols": base_cols,
     "group_options": dummy_energy_group_options,
 }
 
-dummy_demand_parameters = {
+developer_demand_parameters = {
     "page_id": ID_PREFIX,
-    "chart_id": "dum_dem",
-    "sec_id": "dum-dem",
+    "chart_id": "dev_dem",
+    "sec_id": "dev-dem",
     "filters": dummy_demand_filters,
     "section_title": "Infeasible service demand",
     "base_cols": base_cols,
     "group_options": dummy_demand_group_options,
 }
 
+technology_capacity_parameters = {
+    "page_id": ID_PREFIX,
+    "chart_id": "technology_capacity",
+    "sec_id": "technology-capacity",
+    "filters": technology_capacity_filters,
+    "section_title": "Technology capacity",
+    "base_cols": base_cols,
+    "group_options": technology_capacity_group_options,
+}
+
 # SERVER ------------------------------------------------------------------
 
 
-def dummy_server(inputs, outputs, session, selected_scens):
+def developers_server(inputs, outputs, session, selected_scens):
     """
     Server functions for dummy process module
     """
@@ -135,8 +183,8 @@ def dummy_server(inputs, outputs, session, selected_scens):
         return tuple(selected_scens["scenario_list"]())
 
     register_server_functions_for_explorer(
-        dummy_demand_parameters,
-        get_dum_demand_df,
+        developer_demand_parameters,
+        get_dev_demand_df,
         scen_tuple,
         selected_scens["is_comparison"],
         inputs,
@@ -145,8 +193,18 @@ def dummy_server(inputs, outputs, session, selected_scens):
     )
 
     register_server_functions_for_explorer(
-        dummy_energy_parameters,
-        get_dum_energy_df,
+        developer_energy_parameters,
+        get_dev_energy_df,
+        scen_tuple,
+        selected_scens["is_comparison"],
+        inputs,
+        outputs,
+        session,
+    )
+
+    register_server_functions_for_explorer(
+        technology_capacity_parameters,
+        get_base_technology_capacity_df,
         scen_tuple,
         selected_scens["is_comparison"],
         inputs,
@@ -158,11 +216,15 @@ def dummy_server(inputs, outputs, session, selected_scens):
 # UI --------------------------------------------------
 
 
-sections = [dummy_demand_parameters, dummy_energy_parameters]
+sections = [
+    developer_demand_parameters,
+    developer_energy_parameters,
+    technology_capacity_parameters,
+]
 
 
-dummy_ui = make_explorer_page_ui(
+developers_ui = make_explorer_page_ui(
     sections,
     ID_PREFIX,
-    page_info_button_id="info_dum",
+    page_info_button_id="info_dev",
 )
