+++ TIMES-NZ-INTERNAL-QA/src/times_nz_internal_qa/app/app_module_utilities.py
@@ -20,7 +20,7 @@ readme_app_use = (ASSETS / "app_use.md").read_text(encoding="utf-8")
 
 info_pri_doc = (ASSETS / "docs/primary_energy.md").read_text(encoding="utf-8")
 info_elc_doc = (ASSETS / "docs/electricity_generation.md").read_text(encoding="utf-8")
-info_dum_doc = (ASSETS / "docs/infeasibilities.md").read_text(encoding="utf-8")
+info_dev_doc = (ASSETS / "docs/infeasibilities.md").read_text(encoding="utf-8")
 info_ems_doc = (ASSETS / "docs/emissions.md").read_text(encoding="utf-8")
 info_dem_doc = (ASSETS / "docs/energy_demand.md").read_text(encoding="utf-8")
 info_esd_doc = (ASSETS / "docs/energy_service_demand.md").read_text(encoding="utf-8")
@@ -59,7 +59,7 @@ def info_server(inputs, outputs, session):
     attach_info("info_ems", info_ems_doc, "Emissions")
     attach_info("info_dem", info_dem_doc, "Energy demand")
     attach_info("info_esd", info_esd_doc, "Energy service demand")
-    attach_info("info_dum", info_dum_doc, "Infeasibilities")
+    attach_info("info_dev", info_dev_doc, "For developers")
 
     # full results download zip
     @render.download(filename="times_nz_3_wip_all_results.zip")
